+++ resources/views/desktop/backend/pages.blade.php
@@ -6,7 +6,7 @@
 
 @section('styles')
 <style>
-    .blog-item > .active::after {
+    .blog-item>.active::after {
         content: url("../images/icons8-edit-48@2x.png");
         width: 24px;
         height: 24px;
@@ -15,12 +15,12 @@
         /* margin-top: -5px; */
     }
 
-    .bk-active::after{
+    .bk-active::after {
         content: url("../images/icons8-tick-box-48@2x.png");
         width: 24px;
         height: 24px;
         right: 0;
-        top:0;
+        top: 0;
         position: absolute;
         margin-left: 5px;
     }
@@ -29,8 +29,8 @@
         position: relative;
     }
 
-    #imagePreview {
-        width: 122px;
+    .preview {
+        /* width: 122px; */
         height: 122px;
     }
 
@@ -122,15 +122,53 @@
                 <p class="font_26 mb-2">
                     最新優惠
                 </p>
-                <div class="grid grid-cols-2 col-gap-4 row-gap-2 p-4">
-                    <label><input class="mr-3" type="checkbox" />優惠1</label>
-                    <label><input class="mr-3" type="checkbox" />優惠2</label>
-                    <label><input class="mr-3" type="checkbox" />優惠3</label>
-                    <label><input class="mr-3" type="checkbox" />優惠4</label>
-                    <label><input class="mr-3" type="checkbox" />優惠5</label>
-                    <label><input class="mr-3" type="checkbox" />優惠6</label>
-                    <label><input class="mr-3" type="checkbox" />優惠7</label>
-                    <label><input class="mr-3" type="checkbox" />優惠8</label>
+                <div class="grid grid-cols-2 col-gap-4 row-gap-2 p-4" id="promotion-wrapper">
+                    <?php
+                    $promotions = App\Blog::where('used_promotion', true)->where('as_promotion', true)->orderBy('column', 'asc')->get();
+                    $data = array_fill(1, 6, null);
+                    foreach ($promotions as $promotion) {
+                        $data[$promotion->column] = $promotion;
+                    }
+                    ?>
+                    @foreach($data as $key => $item)
+                    <label class="promotion-item"><input {{$item ? '' : 'disabled'}} {{$item && $item->state ? 'checked' : ''}} data-id="{{$item ? $item->id : ''}}" class="checkbox mr-3" type="checkbox" />
+                        {{$key.'. '}}<span>{{$item ? $item->title : '"Title of the news"'}}</span>
+                    </label>
+                    @endforeach
+                </div>
+                <div class="flex mx-2">
+                    <div class="w-1/3 mr-12">
+                        <p class="mb-2">Select Position</p>
+                        <div class="w-full inline-block relative">
+                            <select id="position-select" class="block appearance-none w-full bg-white border border-gray-200 px-4 py-2 pr-8 leading-tight focus:outline-none">
+                                <option value="" selected disabled class="text-grey">Position</option>
+                                @for($i = 1; $i < 7; $i++) <option value="{{$i}}">{{$i}}</option>
+                                    @endfor
+                            </select>
+                            <div class="pointer-events-none absolute inset-y-0 right-0 flex items-center px-2 text-gray-700">
+                                <svg class="fill-current h-6 w-8" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
+                                    <path d="M9.293 12.95l.707.707L15.657 8l-1.414-1.414L10 10.828 5.757 6.586 4.343 8z" /></svg>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="w-2/3">
+                        <p class="mb-2">Select news</p>
+                        <div class="w-full inline-block relative">
+                            <select id="news-select" class="block appearance-none w-full bg-white border border-gray-200 px-4 py-2 pr-8 leading-tight focus:outline-none">
+                                <option value="" selected disabled class="text-grey">News</option>
+                                <?php
+                                $promotionNews = App\Blog::where('used_promotion', true)->get();
+                                ?>
+                                @foreach($promotionNews as $news)
+                                <option value="{{$news->id}}" class="">{{$news->title}}</option>
+                                @endforeach
+                            </select>
+                            <div class="pointer-events-none absolute inset-y-0 right-0 flex items-center px-2 text-gray-700">
+                                <svg class="fill-current h-6 w-8" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
+                                    <path d="M9.293 12.95l.707.707L15.657 8l-1.414-1.414L10 10.828 5.757 6.586 4.343 8z" /></svg>
+                            </div>
+                        </div>
+                    </div>
                 </div>
             </div>
         </div>
@@ -147,7 +185,7 @@
                 </div>
                 <div class="flex justify-between">
                     <button type="submit" class="font_21 px-8 py-1 text-white mr-4" style="background-color: #B881FD;">儲存</button>
-                    <button type="button" id="deleteBtn" class="font_21 px-8 py-1 text-white mr-4" style="background-color: #FD8181;">刪除</button>
+                    <button type="button" class="deleteBtn font_21 px-8 py-1 text-white mr-4" style="background-color: #FD8181;">刪除</button>
                     <button type="reset" id="resetBtn" class="font_21 px-4 py-1 text-black">取消</button>
                 </div>
             </div>
@@ -157,16 +195,35 @@
                     <p class="font_26 mb-2">
                         圖片
                     </p>
-                    <p class="font_16 robert-regular mb-2" id="imageUrl"></p>
-                    <div class="flex mb-4">
-                        <img class="mr-4" id="imagePreview" width="122" height="122" />
-                        <div class="flex flex-col justify-between ml-20">
-                            <div>
-                                <button id="inputBtn" type="button" class="appearance-none h-7 w-24 font_14 px-3 align-middle rounded border border-third g-grey-2">選擇圖片</button>
-                                <input type="file" id="imageInput" name="image" class="hidden" accept=".jpg,.png,.gif" />
-                            </div>
+                    <p class="font_16 robert-regular mb-2">Content Picture</p>
+                    <div class="flex w-1/2 mb-4 image-wrapper">
+                        <img id="imagePreview" class="mr-4 preview w-1/2" />
+                        <div class="w-1/2 flex flex-col justify-center">
+                            <img class="object-fill mx-auto cursor-pointer input-wrapper" src="{{asset('images/icons8-plus-64@2x.png')}}" />
+                            <input type="file" name="image" class="hidden image-input" accept=".jpg,.png,.gif" />
+                        </div>
 
-                            <label class="robert-regular"><input class="mr-3" type="checkbox" name="usedNotiBar" id="usedNotiBar" />Use as notification bar</label>
+                    </div>
+                    <div class="flex">
+                        <div class="w-1/2">
+                            <p class="font_16 robert-regular mb-2">Thumbnail Picture</p>
+                            <div class="flex mb-4 image-wrapper">
+                                <img id="thumbPreview" class="mr-4 preview w-1/2" />
+                                <div class="w-1/2 flex flex-col justify-center">
+                                    <img class="object-fill mx-auto cursor-pointer input-wrapper" src="{{asset('images/icons8-plus-64@2x.png')}}" />
+                                    <input type="file" name="thumbnail" class="hidden image-input" accept=".jpg,.png,.gif" />
+                                </div>
+                            </div>
+                        </div>
+                        <div class="w-1/2">
+                            <label class="robert-regular"><input class="mb-4 mr-3" type="checkbox" name="usedPromition" id="usedPromition" />Use as front page promotion</label>
+                            <div class="flex mb-4 image-wrapper">
+                                <img id="promoPreview" class="mr-4 preview w-1/2" />
+                                <div class="w-1/2 flex flex-col justify-center">
+                                    <img class="object-fill mx-auto cursor-pointer input-wrapper" src="{{asset('images/icons8-plus-64@2x.png')}}" />
+                                    <input type="file" name="promotion" class="hidden image-input" accept=".jpg,.png,.gif" />
+                                </div>
+                            </div>
                         </div>
                     </div>
                     <div class="border-b border-gray-300 mb-4">
@@ -187,9 +244,16 @@
                         $blogs = App\Blog::orderBy('updated_at', 'desc')->paginate(5);
                         ?>
                         @foreach ($blogs as $blog)
-                        <div class="mb-2 robert-regular cursor-pointer blog-item" id="{{$blog->id}}">
+                        <div class="flex justify-between mb-2 robert-regular cursor-pointer blog-item" id="{{$blog->id}}">
                             <span class="blog-title">{{ $blog->title }}</span>
-                            <span class="float-right font_12">{{$blog->updated_at->format('d/m/Y')}}</span>
+                            <div class="flex justify-end flex-shrink-0">
+                                <div class="action-bar hidden">
+                                    <img src="{{asset('images/icons8-edit-48@2x.png')}}" class="inline" />
+                                    <span class="deleteBtn ml-2">delete</span>
+                                </div>
+                                <span class="ml-4 float-right font_12 my-auto">{{$blog->updated_at->format('d/m/Y')}}</span>
+                            </div>
+
                         </div>
                         @endforeach
 
@@ -229,48 +293,49 @@
         $("#backgroundForm").submit();
     });
 
-    $("#imageInput").change(function() {
-        $("#imageUrl").text($(this).val());
-        readURL(this, $('#imagePreview'));
-    });
-
-    $("#inputBtn").click(function() {
-        $("#imageInput").click();
-    })
-
     $("#bkAddBtn").click(function() {
         $("#bkInput").click();
     })
 
     $("#resetBtn").click(function() {
-        $(".blog-title").removeClass("active");
-        $('#imagePreview').attr('src', null);
+        $(".action-bar").hide();
+        $('#blogForm').find('.preview').attr('src', null);
         $("#blogId").val(0);
+        $("#method").val('POST');
     })
 
-    $("#deleteBtn").click(function() {
-
+    $(".deleteBtn").click(function(e) {
+        e.preventDefault();
         const blogId = $("#blogId").val();
 
         if (blogId <= 0) {
             return;
         }
 
-        $.ajax({
-            url: '/blog/' + blogId,
-            type: 'DELETE',
-            datatype: 'json',
-            success: function(result) {
-                if (result) {
-                    window.location.reload();   
+        if (confirm("Do you really want to delete this blog?")) {
+            $.ajax({
+                url: '/blog/' + blogId,
+                type: 'DELETE',
+                datatype: 'json',
+                success: function(result) {
+                    if (result) {
+                        window.location.reload();
+                    }
                 }
-            }
-        });
+            });
+        }
     });
 
     $(".blog-item").click(function() {
-        $(".blog-title").removeClass("active");
-        $(this).find('.blog-title').addClass("active");
+        // $(".blog-title").removeClass("active");
+        // $(this).find('.blog-title').addClass("active");
+        if ($(this).hasClass('active')) {
+            return;
+        }
+        $('.blog-item').removeClass('active');
+        $(this).addClass('active');
+        $('.action-bar').hide();
+        const actionBar = $(this).find('.action-bar');
         const blogId = $(this).attr('id');
         //load blog data 
         $.ajax({
@@ -278,23 +343,79 @@
             type: 'GET',
             datatype: 'json',
             success: function(blog) {
-                if (blog.image == null) {
-                    $('#imagePreview').attr('src', null);
-                } else {
-                    $('#imagePreview').attr('src', blog.image);
-                }
-                $("#usedNotiBar").prop("checked", blog.used_notibar);
+                $('#imagePreview').attr('src', blog.image);
+                $('#thumbPreview').attr('src', blog.thumbnail);
+                $('#promoPreview').attr('src', blog.promotion);
+
+                $("#usedPromition").prop("checked", blog.used_promotion);
                 $("#title").val(blog.title);
                 $("#content").val(blog.content);
                 $("#method").val('PUT');
                 $("#blogId").val(blog.id);
+
+                actionBar.show();
             }
         });
+    })
+
+    $(".input-wrapper").click(function() {
+        $(this).next(".image-input").click();
+    })
+
+    $(".image-input").change(function() {
+        readURL(this, $(this).closest('.image-wrapper').find('.preview'));
+    });
+
+    $("#position-select").change(function() {
+        $("#news-select").val('');
+    })
+
+    $(".checkbox").change(function() {
+        const id = Number($(this).attr('data-id'));
+        if (Number.isNaN(id)) {
+            return;
+        }
+
+        $.ajax({
+            url: '/blog/set-promotion/' + id + '/' + (this.checked ? 1 : 0),
+            type: 'GET',
+            success: function(result) {
+                alert(result ? "ok" : "fail");
+            }
+        });
+    })
+
+    $("#news-select").change(function() {
+        const position = $("#position-select").val();
+        const news = $(this).val();
+        //check duplication
+        if (position && news) {
+            if ($("[data-id='" + news + "']").length) {
+                alert("Duplicate");
+                return;
+            }
+            $.ajax({
+                url: '/blog/as-promotion/' + news + '/' + position,
+                type: 'GET',
+                success: function(result) {
+                    alert(result ? "ok" : "fail");
+                    const item = $(".promotion-item").eq(position - 1);
+                    const checkbox = item.find('input');
+                    checkbox.removeAttr("checked");
+                    checkbox.removeAttr("disabled");
+                    checkbox.attr('data-id', news);
+                    item.find('span').text($("#news-select option:selected").text());
+                }
+            });
+
+        }
     })
 
 
     $(function() {
         $("#blogId").val(0);
+        $("#position-select").val('');
+        $("#news-select").val('');
         $("#resetBtn").click();
     });
 </script>
